feat(meeting): add meeting exit functionality with proper channel cleanup
- Implemented exit meeting endpoint in MeetingController
- Added MeetingExitObj to handle exit user data and status
- Integrated meeting exit logic in MeetingServiceImpl with redis updates
- Enhanced ChannelContextUtil to remove user contexts from meeting groups
- Added database update support for meeting member status via mapper XML
- Modified MessageSendDto to use Integer messageId instead of Long
- Updated websocket handler to properly pass meetingId as Integer
- Added MeetingMemberStatus enum integration for exit states
- Implemented online member list filtering and empty room cleanup logic

+++ K-Meeting-Backend/src/main/java/xyz/kbws/model/dto/message/MessageSendDto.java
@@ -30,7 +30,7 @@ public class MessageSendDto<T> implements Serializable {
     
     private Long sendTime;
     
-    private Long messageId;
+    private Integer messageId;
     
     private String fileName;
